Create CVApplication.js & method for change CV
Added empty CVApplication.js file which will generate the content
of the CV application.

Wrote a changeCV method for App.js as well, which changes a property
value of the App.js state to the id of the selected application.
This would be used to make sure the content of the correct CV is
generated.